refactor: rename ActionResult.ts to Result.ts
