+++ packages/messenger-internal/src/shared/Result.ts
